Codigo/QID/qid_publisher/requirements.txt
functions-framework==3.*
gspread
pydrive
numpy
pandas
google-cloud-secret-manager